+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.6.8] - 2025-02-05
+### Changed
+- Enhanced payment link expiry logic to use extended period for future installments
+- Improved payment reminder logic to handle near-due payments
+
 ## [1.6.7] - 2025-02-05
 ### Changed
 - Removed "Upcoming Payments" notice from cart and checkout pages